requirements.txt
pandas~=2.2
requests~=2.32
beautifulsoup4~=4.12
google-auth~=2.36
google-api-python-client~=2.152
SQLAlchemy~=2.0
pymysql~=1.1
pytest~=8.0
pytest-cov~=5.0
pytest-mock~=3.12